requirements.txt
python3-kivy
python3-pyaudio
portaudio19-dev
pathlib

python3-imageio-ffmpeg

python3-playsound3
